+++ src/Exception/RegisterUserFailed.php
@@ -16,13 +16,13 @@ class RegisterUserFailed extends Base
     /**
      * @param string $hydroId
      * @param string $message
-     * @param Exception $previousException
+     * @param Exception|null $previousException
      * @return RegisterUserFailed
      */
     public static function withHydroId(
         string $hydroId,
         string $message,
-        ?Exception $previousException = null
+        Exception $previousException = null
     ): RegisterUserFailed {
         return new static(sprintf(
             'Could not register user with Hydro ID %s: %s',