+++ src/Exception/UnableToAcquireAccessToken.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Adrenth\Raindrop\Exception;
+
+/**
+ * Class UnableToAcquireAccessToken
+ *
+ * @package Adrenth\Raindrop\Exception
+ */
+class UnableToAcquireAccessToken extends Base
+{
+
+}