+++ src/Exception/UnregisterUserFailed.php
@@ -16,13 +16,13 @@ class UnregisterUserFailed extends Base
     /**
      * @param string $hydroId
      * @param string $message
-     * @param Exception $previousException
+     * @param Exception|null $previousException
      * @return UnregisterUserFailed
      */
     public static function withHydroId(
         string $hydroId,
         string $message,
-        ?Exception $previousException = null
+        Exception $previousException = null
     ): UnregisterUserFailed {
         return new static(sprintf(
             'Could not unregister user with Hydro ID %s: %s',